feat: release sprint 2 0.3

+++ src/controllers/artist.controller.ts
@@ -19,19 +19,23 @@ export const createArtist = async (req: Request, res: Response) => {
         });
         res.status(201).json(artist);
     } catch (error) {
-        logger.error('error creating artist', error);
+        logger.error('error creating artist' + error);
         res.status(500).json({ message: 'Error creating artist', error });
     }
 };
 
 export const getArtists = async (req: Request, res: Response) => {
+    const { limit = 10, offset = 0 } = req.query;
+
     try {
         const artists: IArtist[] = await prisma.artist.findMany({
             include: { image: true },
+            take: Number(limit),
+            skip: Number(offset),
         });
         res.status(200).json(artists);
     } catch (error) {
-        logger.error('error fetching artists', error);
+        logger.error('error fetching artists' + error);
         res.status(500).json({ message: 'Error fetching artists', error });
     }
 };
@@ -58,7 +62,7 @@ export const getArtistById = async (req: Request, res: Response) => {
             res.status(404).json({ message: 'Artist not found' });
         }
     } catch (error) {
-        logger.error('error fetching artist', error);
+        logger.error('error fetching artist' + error);
         res.status(500).json({ message: 'Error fetching artist', error });
     }
 };
@@ -77,23 +81,38 @@ export const updateArtist = async (req: Request, res: Response) => {
         });
         res.status(200).json(artist);
     } catch (error) {
-        logger.error('error updating artist', error);
+        logger.error('error updating artist' + error);
         res.status(500).json({ message: 'Error updating artist', error });
     }
 };
 
-export const deleteArtist = async (req: Request, res: Response) => {
+export const deleteArtist = async (
+    req: Request,
+    res: Response,
+): Promise<void> => {
     const { id } = req.params;
 
-    if (!id) res.status(400).json({ message: 'Artist ID is required' });
+    if (!id) {
+        res.status(400).json({ message: 'Artist ID is required' });
+        return;
+    }
 
     try {
+        const artistExists = await prisma.artist.findUnique({
+            where: { id: parseInt(id, 10) },
+        });
+
+        if (!artistExists) {
+            res.status(404).json({ message: 'Artist not found' });
+            return;
+        }
+
         await prisma.artist.delete({
             where: { id: parseInt(id, 10) },
         });
         res.status(204).send();
     } catch (error) {
-        logger.error('error deleting artist', error);
+        logger.error('error deleting artist' + error);
         res.status(500).json({ message: 'Error deleting artist', error });
     }
 };
